moved themes to their own tsx file
