+++ Themes.tsx
@@ -1,0 +1,52 @@
+export const themes = {
+  darkBlue: {
+    colors: {
+      primary: '#5c73ff',
+      primaryHover: '#8899ff',
+      pink: 'papayawhip',
+      text: '#b3b9c5',
+      textHighlight: '#fafafa',
+      body: '#202020',
+      darkBackground: '#252525',
+      black: '#0e1111',
+    },
+  },
+
+  lightTheme: {
+    colors: {
+      primary: '#E28DA9',
+      primaryHover: '#EDB6C8',
+      buttonText: '#FAFAFA',
+      buttonTextHover: '#DB7093',
+      buttonBackground: '#FAFA',
+      text: '#b3b9c5',
+      textHighlight: '#FFF',
+      headerText: '#cecece',
+      body: '#FAFAFA',
+      darkBackground: '#252525',
+      black: '#0e1111',
+      codeBackground: '#3A3A3A',
+    },
+  },
+  pinkDarkTheme: {
+    colors: {
+      primary: '#E28DA9',
+      primaryHover: '#EDB6C8',
+      buttonText: '#FAFAFA',
+      buttonTextHover: '#DB7093',
+      buttonBackground: '#181818',
+      text: '#b3b9c5',
+      textHighlight: '#FFF',
+      headerText: '#cecece',
+      body: '#202020',
+      darkBackground: '#252525',
+      black: '#0e1111',
+      codeBackground: '#3A3A3A',
+    },
+  },
+};
+
+/* 
+High-emphasis text -> #FFFFFF (87% opacity)
+Medium-emphasis text -> #FFFFFF (60% opacity)
+Disabled text -> #FFFFFF (38% opacity) */
